Fixed package structure
Fixed package structure. Renamed app packed, app itself to interp/Interp.
