--- app/src/main/java/com/critt/trandroidlator/InterpApp.kt
@@ -1,15 +1,0 @@
-package com.critt.trandroidlator
-
-import android.app.Application
-import com.google.firebase.FirebaseApp
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class InterpApp : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        Timber.plant(Timber.DebugTree())
-        FirebaseApp.initializeApp(this)
-    }
-}
