--- app/src/main/java/com/critt/trandroidlator/ui/AuthActivity.kt
@@ -1,89 +1,0 @@
-package com.critt.trandroidlator.ui
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Toast
-import androidx.activity.viewModels
-import com.critt.trandroidlator.ui.AuthViewModel.AuthState
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import com.firebase.ui.auth.AuthUI
-import com.firebase.ui.auth.FirebaseAuthUIActivityResultContract
-import com.firebase.ui.auth.data.model.FirebaseAuthUIAuthenticationResult
-import com.google.firebase.auth.FirebaseAuth
-import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
-
-@AndroidEntryPoint
-class AuthActivity : AppCompatActivity() {
-    private val signInLauncher = registerForActivityResult(
-        FirebaseAuthUIActivityResultContract()
-    ) { result ->
-        onSignInResult(result)
-    }
-
-    private val viewModel: AuthViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        createSignInIntent()
-
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.toastMessage.collect { message ->
-                    message?.let {
-                        Toast.makeText(this@AuthActivity, message, Toast.LENGTH_SHORT).show()
-                    }
-                }
-            }
-        }
-
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.authState.collect { state ->
-                    when (state) {
-                        AuthState.Authenticated ->
-                            startActivity(
-                                Intent(
-                                    this@AuthActivity,
-                                    MainActivity::class.java
-                                )
-                            )
-
-                        AuthState.Unauthenticated -> {}
-                    }
-                }
-            }
-        }
-    }
-
-    private fun createSignInIntent() {
-        // Choose authentication providers
-        val providers = arrayListOf(
-            AuthUI.IdpConfig.EmailBuilder().setAllowNewAccounts(false).build(),
-        )
-
-        // Create and launch sign-in intent
-        val signInIntent = AuthUI.getInstance()
-            .createSignInIntentBuilder()
-            .setIsSmartLockEnabled(false)
-            .setAvailableProviders(providers)
-            .build()
-        signInLauncher.launch(signInIntent)
-    }
-
-    private fun onSignInResult(result: FirebaseAuthUIAuthenticationResult) {
-        if (result.resultCode == RESULT_OK) {
-            // Successfully signed in
-            FirebaseAuth.getInstance().currentUser?.let {
-                viewModel.initSession(it)
-            }
-        } else {
-            result.idpResponse?.error?.message?.let {
-                viewModel.showToast(it)
-            }
-        }
-    }
-}
