+++ core/src/main/java/com/critt/core/Constants.kt
@@ -1,0 +1,7 @@
+package com.critt.core
+
+object Constants {
+    // Shared Preferences
+    const val PREFS_NAME = "enc_prefs"
+    const val PREFS_KEY_FB_TOKEN = "fb_token"
+}
