--- core/src/main/java/com/critt/trandroidlator/core/Constants.kt
@@ -1,7 +1,0 @@
-package com.critt.trandroidlator.core
-
-object Constants {
-    // Shared Preferences
-    const val PREFS_NAME = "enc_prefs"
-    const val PREFS_KEY_FB_TOKEN = "fb_token"
-}
