--- data/src/main/java/com/critt/trandroidlator/data/di/AudioModule.kt
@@ -1,17 +1,0 @@
-package com.critt.trandroidlator.data.di
-
-import com.critt.trandroidlator.data.AudioSource
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AudioModule {
-    @Provides
-    @Singleton
-    fun provideAudioSource(): AudioSource =
-        AudioSource()
-}
